requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
Brotli==1.0.9
cffi==1.15.1
cssselect2==0.7.0
decorator==5.1.1
Django==4.1.5
django-factory-boy==1.0.0
djangorestframework==3.14.0
executing==1.2.0
factory-boy==3.2.1
Faker==16.6.0
fonttools==4.38.0
html5lib==1.1
ipdb==0.13.11
ipython==8.8.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydyf==0.5.0
Pygments==2.14.0
pyphen==0.13.2
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tinycss2==1.2.1
tomli==2.0.1
traitlets==5.8.1
wcwidth==0.2.6
weasyprint==57.2
webencodings==0.5.1
zopfli==0.2.2
